+++ CHANGELOG.md
@@ -1,6 +1,14 @@
+## [0.1.1] - 8 September 2022
+
+- Fix `SkadiNavigator` popTime
+- Add new BuildContext extension for `SkadiNavigator`
+- Improve `Dot` widget
+- Add more `Examples`
+
 ## [0.1.0] - 7 September 2022
 
 - Fix `SkadiResponsive` value being negative
+- Add new widgets
 
 ## [0.0.3] - 7 September 2022